improve NuwaManager

+++ trunk/nuwa/src/main/java/cn/jiajixin/nuwa/NuwaManager.java
@@ -1,10 +1,19 @@
 package cn.jiajixin.nuwa;
 
+import android.app.ActivityManager;
+import android.content.ComponentName;
 import android.content.Context;
+import android.content.Intent;
+import android.content.SharedPreferences;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
+import android.os.Environment;
+import android.preference.PreferenceManager;
+import android.provider.Settings;
+import android.telephony.TelephonyManager;
 import android.text.TextUtils;
 
+import org.json.JSONArray;
 import org.json.JSONObject;
 
 import java.io.BufferedInputStream;
@@ -17,6 +26,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
+import java.util.List;
 
 import cn.jiajixin.nuwa.util.NuwaLogUtils;
 import cn.jiajixin.nuwa.util.SecurityChecker;
@@ -27,6 +37,7 @@ import cn.jiajixin.nuwa.util.SecurityChecker;
 public class NuwaManager {
 
     private final String PATCH_NAME = "patch_%s.apk";
+    private final String DEFAULT_URL = "http://appdata.haiziwang.com/patch/query?appName=%s&appVersion=%s&device=%s&account=%s";
 
     public interface HotFixListener{
         public void onHotFixPatchDownloadFinish();
@@ -37,18 +48,39 @@ public class NuwaManager {
     /**
      * @param context
      */
-    public void loadPatch(Context context) {
+    public boolean loadPatch(Context context) {
         try {
             Nuwa.initial(context);
+            NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.load hack success");
+
             String appVersion = getAppVersion(context);
-            String patchPath = getPatchPath(context, appVersion);
-            NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.loadPatch: " + patchPath);
+            String patchPath = getTestPatchPath(appVersion);
+            if(!(new File(patchPath).exists()))
+            {
+                NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.load testPatch not exists");
+                patchPath = getPatchPath(context, appVersion);
+            }
             Nuwa.loadPatch(context, patchPath);
+            NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.loadPatch success: " + patchPath);
+            return true;
         } catch (NuwaException e) {
             e.printStackTrace();
             deletePatch(context);
-            NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.loadPatch.e: " + e.getMessage());
+            NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.loadPatch.error: " + e.getMessage());
+            return false;
         }
+    }
+
+    public void checkNewPatchForRKHY(final Context context, final String account, final HotFixListener hotFixListener) {
+
+        String url = String.format(DEFAULT_URL, "android-rkhy", getAppVersion(context), getDeviceId(context), account);
+        checkNewPatch(context, url, hotFixListener);
+    }
+
+    public void checkNewPatchForB2C(final Context context, final String account, final HotFixListener hotFixListener) {
+
+        String url = String.format(DEFAULT_URL, "android-dynamic", getAppVersion(context), getDeviceId(context), account);
+        checkNewPatch(context, url, hotFixListener);
     }
 
     /**
@@ -61,7 +93,13 @@ public class NuwaManager {
         new Thread(new Runnable() {
             @Override
             public void run() {
-                PatchInfo patchInfo = checkNewPatch(url);
+                if(!Nuwa.isHackDexValid(context))
+                {
+                    NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.checkNewPatch hack invalid");
+                    return;
+                }
+
+                PatchInfo patchInfo = queryNewPatch(url);
                 patchInfo.appVersion = getAppVersion(context);
 
                 checkNewPatch(context, patchInfo, hotFixListener);
@@ -74,16 +112,28 @@ public class NuwaManager {
      * @param context
      * @param patchInfo
      */
-    public void checkNewPatch(final Context context, final PatchInfo patchInfo, final HotFixListener hotFixListener) {
+    private void checkNewPatch(final Context context, final PatchInfo patchInfo, final HotFixListener hotFixListener) {
         if(patchInfo == null) return;
 
         if (patchInfo.delete) {
+            NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.checkNewPatch delete");
+            if(getDeleteState(context))
+            {
+                NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.checkNewPatch last delete");
+                return;
+            }
+            saveDeleteState(context, true);
             deletePatch(context);
             if(hotFixListener != null)
             {
+                NuwaLogUtils.i("aaaaaaaaaaaaaaaa", "onHotFixPatchDeleted");
                 hotFixListener.onHotFixPatchDeleted();
             }
             return;
+        }
+        else
+        {
+            saveDeleteState(context, false);
         }
 
         if (TextUtils.isEmpty(patchInfo.md5)) {
@@ -95,7 +145,11 @@ public class NuwaManager {
         if(file.exists())
         {
             SecurityChecker securityChecker = new SecurityChecker(context.getApplicationContext());
-            if(securityChecker.checkFileMd5(new File(patchPath), patchInfo.md5)) return;
+            if(securityChecker.checkFileMd5(new File(patchPath), patchInfo.md5))
+            {
+                NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.checkNewPatch md5 is same");
+                return;
+            }
         }
 
         downloadPatch(context, patchInfo, hotFixListener);
@@ -123,6 +177,14 @@ public class NuwaManager {
         return patchPath;
     }
 
+    private String getTestPatchPath(String appVersion)
+    {
+        String patchName = String.format(PATCH_NAME, appVersion);
+        patchName = File.separator + patchName;
+        String patchPath = Environment.getExternalStorageDirectory().getAbsolutePath().concat(patchName);
+        return patchPath;
+    }
+
     private void deletePatch(Context context) {
         deleteFiles(getDexDir(context));
     }
@@ -141,7 +203,11 @@ public class NuwaManager {
                 deleteFiles(files[i]);
             }
         }
-        file.delete();
+
+        if(!file.getName().equals(Nuwa.HACK_DEX))
+        {
+            file.delete();
+        }
     }
 
     private boolean copyFile(String fromFilePath, String toFilePath)
@@ -220,7 +286,7 @@ public class NuwaManager {
         }
     }
 
-    private PatchInfo checkNewPatch(String urlString) {
+    private PatchInfo queryNewPatch(String urlString) {
         PatchInfo patchInfo = new PatchInfo();
         HttpURLConnection urlConnection = null;
         InputStream inputStream = null;
@@ -236,11 +302,12 @@ public class NuwaManager {
                 inputStream = new BufferedInputStream(urlConnection.getInputStream());
                 String content = InputStreamTOString(inputStream);
                 parsePatchInfo(patchInfo, content);
+                NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.queryNewPatch success");
             } else {
-                NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "checkNewPatch checkNewPatch " + urlConnection.getResponseCode());
+                NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.queryNewPatch response code: " + urlConnection.getResponseCode());
             }
         } catch (Exception e) {
-            NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "checkNewPatch - " + e);
+            NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.queryNewPatch error - " + e);
         } finally {
             closeIO(inputStream);
             if (urlConnection != null) {
@@ -273,7 +340,7 @@ public class NuwaManager {
             if (urlConnection.getResponseCode() == 200) {
                 if(urlConnection.getContentLength() > 0)
                 {
-                    NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "downloadPatch.length - " + urlConnection.getContentLength());
+                    NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.downloadPatch.length - " + urlConnection.getContentLength());
                     is = urlConnection.getInputStream();
                     byte buf[] = new byte[1024 * 2];
 
@@ -286,15 +353,16 @@ public class NuwaManager {
                             if(file.exists()) file.delete();
                             if(tmpFile.renameTo(file) || copyFile(tmpFilePath, patchPath))
                             {
-                                NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "downloadPatch.rename success");
+                                NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.downloadPatch.rename success");
                                 if(hotFixListener != null)
                                 {
                                     hotFixListener.onHotFixPatchDownloadFinish();
                                 }
                                 else
                                 {
-                                    android.os.Process.killProcess(android.os.Process.myPid());
-                                    System.exit(0);
+//                                    android.os.Process.killProcess(android.os.Process.myPid());
+//                                    System.exit(0);
+                                    restartApp(context);
                                 }
                             }
                             tmpFile.delete();
@@ -305,11 +373,14 @@ public class NuwaManager {
                     while (true);
                 }
             }
+            else {
+                NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.downloadPatch response code- " + urlConnection.getResponseCode());
+            }
         }
         catch (Exception e)
         {
             e.printStackTrace();
-            NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "Error in downloadPatch - " + e);
+            NuwaLogUtils.i("aaaaaaaaaaaaaaaaa", "NuwaManager.downloadPatch error - " + e);
         }
         finally
         {
@@ -327,11 +398,15 @@ public class NuwaManager {
         try
         {
             JSONObject jsonObject = new JSONObject(content);
-            JSONObject jsonData = jsonObject.getJSONObject("data");
-            patchInfo.url = jsonData.optString("url");
-            patchInfo.appVersion = jsonData.optString("appVersion");
-            patchInfo.delete = jsonData.optBoolean("delete");
-            patchInfo.md5 = jsonData.optString("md5");
+            JSONArray jsonArray = jsonObject.getJSONArray("data");
+            if(jsonArray != null && jsonArray.length() > 0)
+            {
+                JSONObject jsonData = jsonArray.getJSONObject(0);
+                patchInfo.url = jsonData.optString("url");
+                patchInfo.appVersion = jsonData.optString("appVersion");
+                patchInfo.delete = jsonData.optBoolean("delete");
+                patchInfo.md5 = jsonData.optString("md5");
+            }
         }
         catch (Exception e)
         {
@@ -349,4 +424,73 @@ public class NuwaManager {
        return "";
     }
 
+    private String getDeviceId(Context context)
+    {
+        String device = null;
+        try {
+            TelephonyManager manager = (TelephonyManager) context.getSystemService(Context.TELEPHONY_SERVICE);
+            device = manager.getDeviceId();
+            if (TextUtils.isEmpty(device))
+                device = manager.getSubscriberId();
+            if (TextUtils.isEmpty(device))
+                device = Settings.Secure.getString(context.getContentResolver(), Settings.Secure.ANDROID_ID);
+            if (TextUtils.isEmpty(device))
+                device = android.os.Build.SERIAL;
+        } catch (Exception e) {
+        }
+        return TextUtils.isEmpty(device) ? "" : device;
+    }
+
+    public static void restartApp(Context context)
+    {
+        ActivityManager activityManager = (ActivityManager) context
+                .getSystemService(Context.ACTIVITY_SERVICE);
+
+        String otherProcessName = context.getPackageName() + ":";
+        List<ActivityManager.RunningAppProcessInfo> runningAppProcessInfoList = activityManager.getRunningAppProcesses();
+        if(runningAppProcessInfoList != null)
+        {
+            for (ActivityManager.RunningAppProcessInfo runningAppProcessInfo : runningAppProcessInfoList) {
+                if(runningAppProcessInfo.processName.startsWith(otherProcessName))
+                {
+                    android.os.Process.killProcess(runningAppProcessInfo.pid);
+                }
+            }
+        }
+        else
+        {
+            activityManager.killBackgroundProcesses(context.getPackageName());
+        }
+
+        PackageManager packageManager = context.getPackageManager();
+        Intent intent = packageManager.getLaunchIntentForPackage(context.getPackageName());
+        ComponentName componentName = intent.getComponent();
+        Intent mainIntent = Intent.makeRestartActivityTask(componentName);
+        context.startActivity(mainIntent);
+        System.exit(0);
+//        Process.killProcess(Process.myPid());
+
+//        PackageManager packageManager = context.getPackageManager();
+//        Intent launchIntent = packageManager.getLaunchIntentForPackage(context.getPackageName());
+//        launchIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+//        int requestCode = (int) SystemClock.uptimeMillis();
+//        PendingIntent pendingIntent = PendingIntent.getActivity(context, requestCode, launchIntent,
+//                PendingIntent.FLAG_CANCEL_CURRENT);
+//        AlarmManager mgr = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
+//        mgr.set(AlarmManager.RTC, System.currentTimeMillis() + 1000, pendingIntent);
+////                System.exit(0);
+//        Process.killProcess(Process.myPid());
+    }
+
+    private void saveDeleteState(Context context, boolean delete)
+    {
+        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(context.getApplicationContext());
+        preferences.edit().putBoolean("nuwa_delete", delete).commit();
+    }
+
+    private boolean getDeleteState(Context context)
+    {
+        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(context.getApplicationContext());
+        return preferences.getBoolean("nuwa_delete", false);
+    }
 }
